# 8260718: Reconsider SymbolTable shared/dynamic switch optimization